backend/requirements.txt
flask==2.3.0
flask-cors==4.0.0
flask-socketio==5.3.0
opencv-python==4.8.0
mediapipe==0.10.3
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
python-socketio==5.9.0
deep-translator==1.11.4
python-dotenv==1.0.0